// types.ts
export interface ClientData {
  name: string;
  value: number;
}

export interface SipData {
  name: string;
  uv: number;
  pv: number;
}

export interface MisData {
  name: string;
  uv: number;
  pv: number;
  amt: number;
}

export interface StatData {
    title: string;
    value: string;
    detail: string;
}